put debug window in example helper mod
