+++ src/test_util.rs
@@ -286,56 +286,3 @@ pub mod geometry {
         height_to_triangles(height_map, terrain_res, terrain_res)
     }
 }
-
-pub mod atomic_colors {
-    use glam::{vec4, Vec3, Vec4, Vec4Swizzles};
-    use std::sync::{
-        atomic::{AtomicU32, Ordering},
-        Arc,
-    };
-
-    /// A very basic buffer for async debug rendering
-    #[derive(Clone)]
-    pub struct AtomicColorBuffer {
-        pub data: Arc<Vec<[AtomicU32; 4]>>,
-        pub width: usize,
-        pub height: usize,
-    }
-    impl AtomicColorBuffer {
-        pub fn new(width: usize, height: usize) -> Self {
-            Self {
-                data: Arc::new(
-                    (0..width * height)
-                        .map(|_| [const { AtomicU32::new(0) }; 4])
-                        .collect::<Vec<_>>(),
-                ),
-                width,
-                height,
-            }
-        }
-        pub fn get(&self, i: usize) -> Vec4 {
-            vec4(
-                f32::from_bits(self.data[i][0].load(Ordering::Relaxed)),
-                f32::from_bits(self.data[i][1].load(Ordering::Relaxed)),
-                f32::from_bits(self.data[i][2].load(Ordering::Relaxed)),
-                f32::from_bits(self.data[i][3].load(Ordering::Relaxed)),
-            )
-        }
-        pub fn set(&self, i: usize, color: Vec4) {
-            self.data[i][0].store(color.x.to_bits(), Ordering::Relaxed);
-            self.data[i][1].store(color.y.to_bits(), Ordering::Relaxed);
-            self.data[i][2].store(color.z.to_bits(), Ordering::Relaxed);
-            self.data[i][3].store(color.w.to_bits(), Ordering::Relaxed);
-        }
-        pub fn get_px(&self, x: usize, y: usize) -> Vec4 {
-            self.get(y * self.width + x)
-        }
-        pub fn set_px(&self, x: usize, y: usize, color: Vec4) {
-            self.set(y * self.width + x, color)
-        }
-    }
-    pub fn color_to_minifb_pixel(color: Vec4) -> u32 {
-        let c = (color.xyz().clamp(Vec3::ZERO, Vec3::ONE) * 255.0).as_uvec3();
-        (c.x << 16) | (c.y << 8) | c.z
-    }
-}
